CI: add malicious uffd page fault handler
