+++ tests/host_tools/uffd/src/lib.rs
@@ -1,0 +1,4 @@
+// Copyright 2022 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod uffd_utils;
